separate two forms
